feat: Add createActivationToken function
Introduce a new function, createActivationToken, to generate activation tokens for user activation. The function utilizes the 'jsonwebtoken' library to create a JWT containing user information and an activation code. The activation code is randomly generated and the JWT is signed with the activation secret from the environment. The token expires after 5 minutes.

Details:
- Added createActivationToken function in [filename]
- Utilizes jwt.sign for token creation
- Generates a random activation code for each token
- Expires the token after 5 minutes